requirements.txt
dash>=2.17.0
dash-bootstrap-components>=1.6.0
databricks-sdk>=0.28.0
databricks-sql-connector>=3.0.0
openai>=1.30.0
plotly>=5.20.0
pandas>=2.0.0
pyarrow>=14.0.0
